+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Method to add a json interface as a string directly, without reading it from a file
+
 ### Changed
 - Bump the Astarte message hub proto definitions to v0.8.0